+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/messenger_routing.php
@@ -15,6 +15,7 @@ $container->loadFromExtension('framework', [
             SecondMessage::class => [
                 'senders' => ['amqp', 'audit'],
             ],
+            'Symfony\*' => 'amqp',
             '*' => 'amqp',
         ],
         'transports' => [