+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/messenger_routing_invalid_wildcard.php
@@ -1,0 +1,17 @@
+<?php
+
+$container->loadFromExtension('framework', [
+    'http_method_override' => false,
+    'serializer' => true,
+    'messenger' => [
+        'serializer' => [
+            'default_serializer' => 'messenger.transport.symfony_serializer',
+        ],
+        'routing' => [
+            'Symfony\*\DummyMessage' => ['audit'],
+        ],
+        'transports' => [
+            'audit' => 'null://',
+        ],
+    ],
+]);