+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
  * Add `DomCrawlerAssertionsTrait::assertSelectorCount(int $count, string $selector)`
  * Allow to avoid `limit` definition in a RateLimiter configuration when using the `no_limit` policy
  * Add `--format` option to the `debug:config` command
+ * Add support to pass namespace wildcard in `framework.messenger.routing`
 
 6.2
 ---